Flask midday update
Tenant register is working better, when a tenant is removed he is also kicked from the room model. im not sure what the login and logout functions do that much of, i may need to look at sessions again. Next i will try to set up a messaging structure that will allow tenants to raise issues with maintenance if need be