Update udpclientserver.py
Rendevous server only connects 2 udpcs at a time, new servers should automatically be added to the existing network upon connection. Figure out how to send data via scripts, if a new block is mined how will it be sent the the network. if a new tx exists how will it be added to the pool. how are txs removed from the pool if they are minted into a block. how is this shown across networks.