BC update: WEBSITE
developing website to interact with BC, may need to take a javascript detour to acheive it though. Wouldnt be bad to learn javascript or ts. All i need is a home page to create the public key, send that public key to the nodes, have them store it and be able to call it. then i need a wallet with a balance that can be called from the BC